登入功能 -ok

+++ app/Http/Controllers/MessageController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Message;
 use Carbon\Carbon;
-
+use Illuminate\Support\Facades\Auth;
 class MessageController extends Controller
 {
     public function index(Request $request)
@@ -18,7 +18,7 @@ class MessageController extends Controller
     public function store(Request $request)
     {
         $Message =new Message;
-        $Message->memberId = 1;
+        $Message->memberId = Auth::id();
         $Message->comment = $request->comment;
         $Message->save();
         return redirect("/message")->with("status","添加留言成功");
@@ -27,22 +27,30 @@ class MessageController extends Controller
     public function edit($commentNo)
     {
         $message = Message::showEditMsg($commentNo);
-        return view("pages/message/edit", compact("message"));
+        if(Auth::id()==$message->id){
+            return view("pages/message/edit", compact("message"));
+        }else{
+            return redirect("/message")->with("status","僅能編輯自己的留言!");
+        }
     }
 
     public function update(Request $request,  $commentNo)
     {
-        $message = Message::find($commentNo);
-        $message->comment = $request->comment;
-        $message->commentTime = Carbon::now();
-        $message->update();
-        return redirect("/message");
+            $message = Message::find($commentNo);
+            $message->comment = $request->comment;
+            $message->commentTime = Carbon::now();
+            $message->update();
+            return redirect("/message");
     }
 
     public function destroy($commentNo)
     {
         $message = Message::find($commentNo);
-        $message->delete();
-        return redirect()->back()->with("status","刪除留言成功");
+        if(Auth::id()==$message->id){
+            $message->delete();
+            return redirect()->back()->with("status","刪除留言成功");
+        }else{
+            return redirect()->back()->with("status","刪除失敗! 僅能刪除自己的留言");
+        }
     }
 }
